--- photosClone/photosClone/MainNavigationViewController.swift
@@ -1,28 +1,0 @@
-//
-//  MainNavigationViewController.swift
-//  photosClone
-//
-//  Created by Nikita Alpatiev on 1/12/23.
-//
-
-import UIKit
-
-class MainNavigationViewController: UINavigationController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // this class
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
