+++ photosClone/photosClone/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
