+++ photosClone/photosClone/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
