+++ photosClone/photosClone/Resources/Assets.xcassets/Contents.json
