+++ photosClone/photosClone/Resources/Base.lproj/LaunchScreen.storyboard
