+++ photosClone/photosClone/Resources/Info.plist
