+++ photosClone/photosClone/Sources/Application/AppDelegate.swift
@@ -1,0 +1,16 @@
+//
+//  AppDelegate.swift
+//  photosClone
+//
+//  Created by Nikita Alpatiev on 1/12/23.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        return true
+    }
+}
