+++ photosClone/photosClone/Sources/Application/SceneDelegate.swift
@@ -1,0 +1,20 @@
+//
+//  SceneDelegate.swift
+//  photosClone
+//
+//  Created by Nikita Alpatiev on 1/12/23.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        window = UIWindow(windowScene: windowScene)
+        window?.rootViewController = MainTabBarViewController()
+        window?.makeKeyAndVisible()
+    }
+}
