+++ photosClone/photosClone/Sources/ViewControllers/Albums/AlbumsViewController.swift
@@ -1,0 +1,15 @@
+//
+//  AlbumsViewController.swift
+//  photosClone
+//
+//  Created by Nikita Alpatiev on 1/12/23.
+//
+
+import UIKit
+
+class AlbumsViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
