+++ photosClone/photosClone/Sources/ViewControllers/AllPhotos/AllPhotosViewController.swift
@@ -1,0 +1,15 @@
+//
+//  AllPhotosViewController.swift
+//  photosClone
+//
+//  Created by Nikita Alpatiev on 1/12/23.
+//
+
+import UIKit
+
+class AllPhotosViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
