+++ photosClone/photosClone/Sources/ViewControllers/ForYou/ForYouViewController.swift
@@ -1,0 +1,15 @@
+//
+//  ForYouViewController.swift
+//  photosClone
+//
+//  Created by Nikita Alpatiev on 1/12/23.
+//
+
+import UIKit
+
+class ForYouViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
