+++ photosClone/photosClone/Sources/ViewControllers/Main/MainNavigationViewController.swift
@@ -1,0 +1,15 @@
+//
+//  MainNavigationViewController.swift
+//  photosClone
+//
+//  Created by Nikita Alpatiev on 1/12/23.
+//
+
+import UIKit
+
+class MainNavigationViewController: UINavigationController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
