+++ photosClone/photosClone/Sources/ViewControllers/Main/MainTabBarViewController.swift
@@ -1,0 +1,16 @@
+//
+//  MainTabBarViewController.swift
+//  photosClone
+//
+//  Created by Nikita Alpatiev on 1/12/23.
+//
+
+import UIKit
+
+class MainTabBarViewController: UITabBarController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .white
+    }
+}
