+++ photosClone/photosClone/Sources/ViewControllers/Search/SearchViewController.swift
@@ -1,0 +1,15 @@
+//
+//  SearchViewController.swift
+//  photosClone
+//
+//  Created by Nikita Alpatiev on 1/12/23.
+//
+
+import UIKit
+
+class SearchViewController: UIViewController {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
